+++ src/screens/HintScreen/HintScreen.css
@@ -1,0 +1,77 @@
+@import url("https://fonts.googleapis.com/css?family=Fira+Code|Rubik&display=swap");
+
+.background {
+  text-align: center;
+  width: 100%;
+  height: 100%;
+  background-color: #012a46;
+  position: relative;
+  display: flex;
+}
+
+.mobile-container {
+  color: #ff7661;
+  font-family: Rubik;
+  display: flex;
+  max-width: 500px;
+  width: 100%;
+  height: 800px;
+  background-color: #012a46;
+  /*position: relative;*/
+  flex-direction: column;
+  margin: auto;
+}
+
+.page-title {
+  color: #ff7661;
+  font-family: Rubik;
+  font-size: 16px;
+  text-align: center;
+  width: 100%;
+  padding-top: 100px;
+}
+
+.daily-word {
+  color: #ff7661;
+  font-family: Rubik;
+  font-size: 26px;
+  text-align: center;
+  width: 100%;
+  padding-top: 60px;
+}
+
+.main-button {
+  background-color: #ff7661;
+  color: #012a46;
+  font-family: Fira Code;
+  font-weight: bold;
+  font-size: 16px;
+  text-align: center;
+  width: 30%;
+  border-radius: 50px;
+  border-color: transparent;
+  margin: auto;
+}
+
+.dh-image-bg {
+  background-color: #25465d;
+  width: 100%;
+  height: 450px;
+  margin: auto;
+  margin-top: 40px;
+  margin-bottom: 0px;
+}
+
+.dh-image {
+  padding: 10%;
+  display: inline-block;
+  margin: auto;
+}
+
+.cls-1 {
+  fill: none;
+  stroke: #ff7661 !important;
+  stroke-linecap: round;
+  stroke-linejoin: round;
+  stroke-width: 2px;
+}
